Wizards are moved to new location
