+++ plugins/core/hu.bme.mit.gamma.gui/META-INF/MANIFEST.MF
@@ -1,0 +1,13 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Gamma Graphical User Interface
+Bundle-SymbolicName: hu.bme.mit.gamma.gui;singleton:=true
+Bundle-Version: 2.8.0.qualifier
+Export-Package: hu.bme.mit.gamma.gui
+Require-Bundle: org.eclipse.jface,
+ org.eclipse.ui.workbench,
+ org.eclipse.equinox.registry,
+ org.eclipse.core.resources,
+ org.eclipse.ui.ide
+Automatic-Module-Name: hu.bme.mit.gamma.gui
+Bundle-RequiredExecutionEnvironment: JavaSE-17
