+++ plugins/core/hu.bme.mit.gamma.gui/build.properties
@@ -1,0 +1,4 @@
+source.. = src/
+output.. = bin/
+bin.includes = META-INF/,\
+               .
