+++ plugins/core/hu.bme.mit.gamma.gui/plugin.xml
@@ -1,0 +1,149 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?eclipse version="3.4"?>
+<plugin>
+   <extension
+         point="org.eclipse.ui.newWizards">
+      <category
+            id="hu.bme.mit.gamma.ui"
+            name="Gamma Wizards">
+      </category>
+      <wizard
+            category="hu.bme.mit.gamma.ui"
+            class="hu.bme.mit.gamma.gui.NewGCLWizard"
+            descriptionImage="icons/gamma-icon-16.png"
+            icon="icons/gamma-icon-16.png"
+            id="hu.bme.mit.gamma.ui.wizards.NewGCLWizard"
+            name="New Gamma Composition Model Wizard"
+            project="false">
+      </wizard>
+      <wizard
+            category="hu.bme.mit.gamma.ui"
+            class="hu.bme.mit.gamma.gui.NewGCLStatechartWizard"
+            icon="icons/gamma-icon-16.png"
+            id="hu.bme.mit.gamma.ui.wizards.NewGCLStatechartWizard"
+            name="New Gamma Statechart Model Wizard"
+            project="false">
+      </wizard>
+      <wizard
+            category="hu.bme.mit.gamma.ui"
+            class="hu.bme.mit.gamma.gui.NewGCLInterfacesWizard"
+            descriptionImage="icons/gamma-icon-16.png"
+            icon="icons/gamma-icon-16.png"
+            id="hu.bme.mit.gamma.ui.wizards.NewGCLInterfacesWizard"
+            name="New Gamma Intercace Model Wizard"
+            project="false">
+      </wizard>
+      <wizard
+            category="hu.bme.mit.gamma.ui"
+            class="hu.bme.mit.gamma.gui.NewGammaPropertyWizard"
+            descriptionImage="icons/gamma-icon-16.png"
+            icon="icons/gamma-icon-16.png"
+            id="hu.bme.mit.gamma.ui.wizards.NewGammaPropertyWizard"
+            name="New Gamma Property Wizard"
+            project="false">
+      </wizard>
+   </extension>
+ <extension
+       point="org.eclipse.ui.perspectiveExtensions">
+    <perspectiveExtension
+          targetID="*">
+       <newWizardShortcut
+             id="hu.bme.mit.gamma.ui.wizards.NewGCLWizard">
+       </newWizardShortcut>
+       <newWizardShortcut
+             id="hu.bme.mit.gamma.ui.wizards.NewGCLStatechartWizard">
+       </newWizardShortcut>
+       <newWizardShortcut
+             id="hu.bme.mit.gamma.ui.wizards.NewGCLInterfaceWizard">
+       </newWizardShortcut>
+       <newWizardShortcut
+             id="hu.bme.mit.gamma.ui.wizards.NewGammaPropertyWizard">
+       </newWizardShortcut>
+    </perspectiveExtension>
+ </extension>
+ <extension
+       point="org.eclipse.ui.navigator.navigatorContent">
+    <commonWizard
+          menuGroupId="hu.bme.mit.gamma"
+          type="new"
+          wizardId="hu.bme.mit.gamma.ui.wizards.NewGCLWizard">
+       <enablement>
+          <or>
+             <adapt
+                   type="org.eclipse.core.resources.IFile">
+             </adapt>
+             <adapt
+                   type="org.eclipse.core.resources.IFolder">
+             </adapt>
+             <adapt
+                   type="org.eclipse.core.resources.IProject">
+             </adapt>
+          </or>
+       </enablement>
+    </commonWizard>
+ </extension>
+ <extension
+       point="org.eclipse.ui.navigator.navigatorContent">
+    <commonWizard
+          menuGroupId="hu.bme.mit.gamma"
+          type="new"
+          wizardId="hu.bme.mit.gamma.ui.wizards.NewGCLStatechartWizard">
+       <enablement>
+          <or>
+             <adapt
+                   type="org.eclipse.core.resources.IFile">
+             </adapt>
+             <adapt
+                   type="org.eclipse.core.resources.IFolder">
+             </adapt>
+             <adapt
+                   type="org.eclipse.core.resources.IProject">
+             </adapt>
+          </or>
+       </enablement>
+    </commonWizard>
+ </extension>
+ <extension
+       point="org.eclipse.ui.navigator.navigatorContent">
+    <commonWizard
+          menuGroupId="hu.bme.mit.gamma"
+          type="new"
+          wizardId="hu.bme.mit.gamma.ui.wizards.NewGCLInterfacesWizard">
+       <enablement>
+          <or>
+             <adapt
+                   type="org.eclipse.core.resources.IFile">
+             </adapt>
+             <adapt
+                   type="org.eclipse.core.resources.IFolder">
+             </adapt>
+             <adapt
+                   type="org.eclipse.core.resources.IProject">
+             </adapt>
+          </or>
+       </enablement>
+    </commonWizard>
+ </extension>
+ <extension
+       point="org.eclipse.ui.navigator.navigatorContent">
+    <commonWizard
+          menuGroupId="hu.bme.mit.gamma"
+          type="new"
+          wizardId="hu.bme.mit.gamma.ui.wizards.NewGammaPropertyWizard">
+       <enablement>
+          <or>
+             <adapt
+                   type="org.eclipse.core.resources.IFile">
+             </adapt>
+             <adapt
+                   type="org.eclipse.core.resources.IFolder">
+             </adapt>
+             <adapt
+                   type="org.eclipse.core.resources.IProject">
+             </adapt>
+          </or>
+       </enablement>
+    </commonWizard>
+ </extension>
+
+</plugin>
