+++ plugins/core/hu.bme.mit.gamma.gui/src/hu/bme/mit/gamma/gui/NewGCLStatechartWizard.java
@@ -8,7 +8,7 @@
  *
  * SPDX-License-Identifier: EPL-1.0
  ********************************************************************************/
-package hu.bme.mit.gamma.ui.wizards;
+package hu.bme.mit.gamma.gui;
 
 import org.eclipse.jface.viewers.IStructuredSelection;
 import org.eclipse.jface.wizard.Wizard;
